Fix callers of generateUserToken/createUserToken after Integer->long signature change
Pass 0L instead of null for token duration to avoid NullPointerException
when Groovy auto-unboxes null to primitive long. Update mock interface
in WebhookServiceSpec to match new method signature.

+++ rundeckapp/src/test/groovy/rundeck/services/WebhookServiceSpec.groovy
@@ -944,7 +944,7 @@ class WebhookServiceSpec extends Specification implements ServiceUnitTest<Webhoo
     }
 
     interface MockApiService {
-        Map generateUserToken(UserAndRolesAuthContext ctx, Integer expiration, String user, Set<String> roles, boolean forceExpiration, AuthTokenType tokenType) throws Exception
+        Map generateUserToken(UserAndRolesAuthContext ctx, long expiration, String user, Set<String> roles, boolean forceExpiration, AuthTokenType tokenType) throws Exception
     }
 
     interface MockStorageService {
